+++ src/globals.cc
@@ -31,6 +31,34 @@ bool RuntimeInitialized = false;
 const int DefaultNumThreads = 0; // Let the ThreadTeam choose.
 int NumThreads;                  // OMP_NUM_THREADS
 
+// Display the LOMP environment
+void displayEnvironmentVariable(const std::string & name) {
+  std::string value;
+  environment::getString(name.c_str(), value, "");
+  if (value.empty()) {
+    printf("  [host] %s: value is not defined\n", name.c_str());
+  }
+  else {
+    printf("  [host] %s='%s'\n", name.c_str(), value.c_str());
+  }
+}
+
+void displayEnvironment(displayVerbosity verbosity) {
+  auto standardDisplay = {"OMP_NUM_THREADS","OMP_SCHEDULE", "OMP_DISPLAY_ENV"};
+  auto verboseDisplay = {"LOMP_LOCK_KIND", "LOMP_BARRIER_KIND", "LOMP_DEBUG", "LOMP_TRACE"};
+  printf("OPENMP DISPLAY ENVIRONMENT\n");
+  printf("  _OPENMP='%d'\n", 0);
+  for (auto var : standardDisplay) {
+    displayEnvironmentVariable(var);
+  }
+  if (verbosity == displayVerbosity::verbose) {
+    for (auto var : verboseDisplay) {
+      displayEnvironmentVariable(var);
+    }
+  }
+  printf("OPENMP DISPLAY ENVIRONMENT END\n");
+}
+
 // Set up necessary runtime data structures.
 void initializeRuntime() {
   debug(Debug::Announce,
@@ -65,6 +93,20 @@ void initializeRuntime() {
   // and the loop scheduling parse OMP_SCHEDULE so that schedule(runtime)
   // works.
   initializeLoops();
+
+  // Print LOMP environment if requested via OMP_DISPLAY_ENV
+  std::string env;
+  auto verbosity = displayVerbosity::disabled;
+  environment::getString("OMP_DISPLAY_ENV", env, "false");
+  if (env == "true" || env == "1") {
+    verbosity = displayVerbosity::enabled;
+  }
+  else if (env == "verbose") {
+    verbosity = displayVerbosity::verbose;
+  }
+  if (verbosity != displayVerbosity::disabled) {
+    displayEnvironment(verbosity);
+  }
 
   RuntimeInitialized = true;
 }